refactor: distribution map
